Feat/user dto controllers implementation

+++ Controllers/berryBushController.cs
@@ -4,10 +4,13 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Models;
 using DTO;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace Controllers
 {
     [ApiController]
+    [Authorize]
     [Route("api/[controller]")]
     public class BerryBushController : ControllerBase
     {
@@ -21,7 +24,14 @@ namespace Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<BerryBush>>> GetAllBerryBushes()
         {
-            var berryBushes = await _db.BerryBushes.ToListAsync();
+            string userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (userIdClaim == null || !int.TryParse(userIdClaim, out int userId))
+            {
+                return Unauthorized("User Id is not correct");
+            }
+
+            var berryBushes = await _db.BerryBushes.Where(bush => bush.UserId == userId).ToListAsync();
             return Ok(berryBushes);
         }
 
@@ -38,7 +48,9 @@ namespace Controllers
         [HttpPost]
         public async Task<ActionResult<BerryBush>> PostBerryBush(BerryBushCreation berryBush)
         {
-            BerryBush fullBush = new BerryBush(berryBush.Name, berryBush.Species, berryBush.PlantedAt, berryBush.TrellisNeeds);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            BerryBush fullBush = new BerryBush(berryBush.Name, berryBush.Species, berryBush.PlantedAt, berryBush.TrellisNeeds, Convert.ToInt32(userId));
             _db.BerryBushes.Add(fullBush);
             await _db.SaveChangesAsync();
             return Ok(fullBush);
@@ -47,7 +59,20 @@ namespace Controllers
         [HttpDelete]
         public async Task<ActionResult<BerryBush>> DeleteBerryBush(int Id)
         {
+            string userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (userIdClaim == null || !int.TryParse(userIdClaim, out int userId))
+            {
+                return Unauthorized("User Id is not correct");
+            }
+
             BerryBush bush = _db.BerryBushes.Find(Id);
+
+            if (userId != bush.UserId)
+            {
+                return BadRequest("Not acceptable");
+            }
+            
             _db.BerryBushes.Remove(bush);
             await _db.SaveChangesAsync();
             return Ok(bush);
